Ismael Flores iaguilar@quarksoft.net * Se realizan correcciones y modificaciones para perimitir numeros mas grandes en ids de transaccion y folio partida, se hacen validaciones de tamanos en servicio de pagos y se establecen valores de comparacion en calse de constantes de pagos

+++ src/main/java/mx/com/nmp/pagos/mimonte/dao/PagoRepository.java
@@ -33,6 +33,6 @@ public interface PagoRepository extends JpaRepository<Pago, Integer> {
 	 * @throws SQLException
 	 */
 	@Query("SELECT COUNT(p.id) FROM Pago p WHERE p.idTransaccionMidas = :idTransaccionMidas")
-	public Integer countByIdTransaccionMidas(@Param("idTransaccionMidas") Integer idTransaccionMidas)
+	public Integer countByIdTransaccionMidas(@Param("idTransaccionMidas") Long idTransaccionMidas)
 			throws DataIntegrityViolationException, SQLDataException, SQLException;
 }
